Created module directory and module file
