+++ index.html
@@ -5,7 +5,7 @@
  <meta charset="UTF-8">
  <meta http-equiv="X-UA-Compatible" content="IE=edge">
  <meta name="viewport" content="width=device-width, initial-scale=1.0">
- <link rel="stylesheet" href="./style.css" />
+ <link rel="stylesheet" href="./index.css"/>
  <title>Awesome Books</title>
 </head>
 
@@ -67,7 +67,7 @@
   <p>Copyright &copy; <span>Furkan Salim Hıdır, Ben Kiarie &amp; Said Rašinlić</span></p>
  </footer>
 
- <script src="./script.js"></script>
+ <script src="./index.js"></script>
  <script src="https://code.jquery.com/jquery-3.6.0.js" integrity="sha256-H+K7U5CnXl1h5ywQfKtSj8PCmoN9aaq30gDh27Xc0jk=" crossorigin="anonymous"></script>
 </body>
