+++ index.js
@@ -1,0 +1,40 @@
+import Book from './modules/BookClass.js';
+
+Book.updateUi();
+
+Book.listBtn.forEach((btn, i) => {
+  btn.onclick = () => {
+    Book.listBtn.forEach((oldBtn) => {
+      oldBtn.classList.remove('active');
+    });
+    btn.classList.add('active');
+    Book.addActive.forEach((sec, index) => {
+      if (i === index) {
+        sec.classList.add('active');
+      } else {
+        sec.classList.remove('active');
+      }
+    });
+  };
+});
+
+Book.formBtn.addEventListener('click', Book.addBooks);
+
+const timeBox = document.querySelector('#date');
+
+function time() {
+  const date = new Date();
+  const locale = navigator.language;
+  const options = {
+    month: 'long',
+    day: 'numeric',
+    year: 'numeric',
+    hour: 'numeric',
+    minute: 'numeric',
+    second: 'numeric',
+    hour12: 'false',
+  };
+  timeBox.textContent = `${date.toLocaleTimeString(locale, options)}`;
+}
+
+setInterval(time, 1000);
